updatedeps: readd latest packages
