+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - bugfix: When there is no placeholder in the template passed to String.Format it should return the original template
 
 #### 1.5.0
-- feature: !IMPORTANT methods/properties starting with uppercase are marked deprecated now and will be removed with version 2.0.0! Check the deprecated infos
+- Using latest typescript version, please check if your environment still works with this version before upgrading!
+- feature: !IMPORTANT methods/properties starting with uppercase are marked deprecated now and will be removed in future versions! Check the deprecated infos
 - feature: added class `$String` for everybody who is facing issues when `String` is used.
 - feature: `String.isNullOrWhiteSpace` accepts null or undefined now
 - chore: Updated packages
