+++ package.json
@@ -29,16 +29,13 @@
   "homepage": "https://github.com/sevensc/typescript-string-operations#readme",
   "devDependencies": {
     "@types/chai": "^4.2.7",
-    "@types/mocha": "^2.2.48",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^18.7.11",
     "chai": "^4.2.0",
-    "gulp": "^4.0.2",
-    "gulp-rename": "^1.4.0",
-    "gulp-typescript": "^3.2.4",
-    "gulp-uglify": "^3.0.2",
-    "mocha": "^5.2.0",
-    "pump": "^2.0.1",
-    "ts-node": "^3.2.0",
-    "typescript": "^2.9.2"
+    "mocha": "^10.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4",
+    "uglify-js": "^3.17.0"
   },
   "directories": {
     "test": "tests"
