+++ string.d.ts
@@ -8,19 +8,19 @@ export declare class $String {
 
 export declare class String extends $String {
     /**
-     * @deprecated The property should not be used, and will be removed with version 2.0.0 use `String.empty` instead.
+     * @deprecated The property should not be used, and will be removed in future versions! Use `String.empty` instead.
     */
     static Empty: string;
     /**
-     * @deprecated The method should not be used, and will be removed with version 2.0.0 use `String.isNullOrWhiteSpace()` instead.
+     * @deprecated The method should not be used, and will be removed in future versions! Use `String.isNullOrWhiteSpace()` instead.
     */
     static IsNullOrWhiteSpace(value: string | null | undefined): boolean;
     /**
-     * @deprecated The method should not be used, and will be removed with version 2.0.0 use `String.join()` instead.
+     * @deprecated The method should not be used, and will be removed in future versions! Use `String.join()` instead.
     */
     static Join(delimiter: string, ...args: (string | object | Array<any>)[]): string;
     /**
-     * @deprecated The method should not be used, and will be removed with version 2.0.0 use `String.format()` instead.
+     * @deprecated The method should not be used, and will be removed in future version!s Use `String.format()` instead.
     */
     static Format(format: string, ...args: any[]): string;
 }
